+++ tests/conftest.py
@@ -93,14 +93,16 @@ class CheckerTestCase(pylint.testutils.CheckerTestCase):
     def extract_node(self, s):
         return astroid.extract_node(s)
 
-    def _make_message(self, msg_id, node, args):
+    def make_message(self, msg_id, node, args):
         return pylint.testutils.Message(msg_id, node=node, args=args)
 
-    undefined_attribute_msg = lambda self, node, *args: self._make_message('protobuf-undefined-attribute', node, args)
-    enum_value_msg = lambda self, node, *args: self._make_message('protobuf-enum-value', node, args)
-    type_error_msg = lambda self, node, *args: self._make_message('protobuf-type-error', node, args)
-    no_posargs_msg = lambda self, node: self._make_message('protobuf-no-posargs', node, args=None)
-    no_assignment_msg = lambda self, node, *args: self._make_message('protobuf-no-assignment', node, args)
+    undefined_attribute_msg = lambda self, node, *args: self.make_message('protobuf-undefined-attribute', node, args)
+    enum_value_msg = lambda self, node, *args: self.make_message('protobuf-enum-value', node, args)
+    type_error_msg = lambda self, node, *args: self.make_message('protobuf-type-error', node, args)
+    no_posargs_msg = lambda self, node: self.make_message('protobuf-no-posargs', node, args=None)
+    no_assignment_msg = lambda self, node, *args: self.make_message('protobuf-no-assignment', node, args)
+    no_repeated_membership_msg = lambda self, node: self.make_message('protobuf-no-repeated-membership', node, args=None)
+    no_proto3_membership_msg = lambda self, node, *args: self.make_message('protobuf-no-proto3-membership', node, args)
 
 
 @pytest.fixture